Implement user interaction (ongoing)
